Remove unused style module, tag and style
Since this is for quick prototypes, the styles are just going to be
inserted in a style tag that's generated through webpack. I'll configure
this later to use a separate stylesheet.

I also removed an unnecessary TS module and img style so I'm not
scratching my head down the road asking "why did I add this? Do I need
this? At this point, I'm too afraid to ask."